requirements.txt
# Paper Radar dependencies
arxiv>=2.1.0
requests>=2.31.0
PyMuPDF>=1.23.0
anthropic>=0.40.0
openai>=1.0.0
sentence-transformers>=2.2.0
umap-learn>=0.5.0
hdbscan>=0.8.33
chromadb>=0.4.0
gitpython>=3.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
sse-starlette>=1.8.0
pyyaml>=6.0.0
python-multipart>=0.0.6
